Fixed the error for #7.
Couldn't figure out how to order by dept_id using line 63, so did it using lines 64-66. I'm not sure if that's what they're looking for, so if y'all can come up w/ something else, if it's even necessary, that'd be awesome.

Added the html and javascript for part 2, but did not start it